% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRAHDLTP-package.R
\docType{package}
\name{NRAHDLTP-package}
\alias{NRAHDLTP}
\alias{NRAHDLTP-package}
\title{NRAHDLTP: Location Tests for High-Dimensional Problems Including Normal-Reference Approach}
\description{
Provides various classical tests and latest normal-reference approach tests for comparing high-dimensional mean vectors in two and k sample populations.
}
\author{
\strong{Maintainer}: Pengfei Wang \email{nie23.wp8738@e.ntu.edu.sg}

Authors:
\itemize{
  \item Shuqi Luo \email{nie23.ls4909@e.ntu.edu.sg}
  \item Tianming Zhu \email{tianming.zhu@nie.edu.sg}
  \item Bu Zhou \email{bu.zhou@u.nus.edu}
}

}
\keyword{internal}
